+++ bindings/python/addtags.py
@@ -1,5 +1,5 @@
 # Dependencies:
-# - auditwheel>=5,<6
+# - auditwheel>=6,<7
 # Requires AUDITWHEEL_PLAT to be set (e.g. manylinux2014_x86_64)
 # Usage:
 # $ python addtags.py WHEEL_PATH WHEEL_DIR
@@ -8,7 +8,7 @@ from os.path import abspath, basename, exists, isfile
 from os.path import join as pjoin
 from sys import argv
 
-from auditwheel.policy import get_priority_by_name, get_replace_platforms
+from auditwheel.policy import WheelPolicies, get_replace_platforms
 from auditwheel.wheel_abi import analyze_wheel_abi
 from auditwheel.wheeltools import InWheelCtx, add_platforms
 
@@ -29,12 +29,13 @@ def main(wheel_path, abi, wheel_dir):
     if not exists(wheel_dir):
         os.makedirs(wheel_dir)
 
-    reqd_tag = get_priority_by_name(abi)
+    policies = WheelPolicies()
+    reqd_tag = policies.get_priority_by_name(abi)
     out_wheel = repair_wheel(wheel_path, abi, wheel_dir)
 
     if out_wheel is not None:
-        analyzed_tag = analyze_wheel_abi(out_wheel).overall_tag
-        if reqd_tag < get_priority_by_name(analyzed_tag):
+        analyzed_tag = analyze_wheel_abi(policies, out_wheel, frozenset()).overall_tag
+        if reqd_tag < policies.get_priority_by_name(analyzed_tag):
             print(
                 "Wheel is eligible for a higher priority tag. "
                 f"You requested {abi} but I have found this wheel is "